1. Page Seperation 2. login page added
