+++ web/package.json
@@ -19,6 +19,7 @@
     "next-themes": "^0.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.13.0",
     "typescript": "5.1.3"
   }
 }
