+++ web/src/app/common.tsx
@@ -1,0 +1,15 @@
+export const G_SOCIAL_MEDIA_LINKS = [
+    {
+      name: "Instagram",
+      url: "",
+    },
+    {
+      name: "LinkedIn",
+      url: "",
+    },
+    {
+      name: "GitHub",
+      url: "",
+    },
+  ];
+
