+++ web/src/app/page.tsx
@@ -1,11 +1,42 @@
-'use client'
-import { Card5 } from './home'
+"use client";
+import { BrowserRouter } from "react-router-dom";
+import { NextUIProvider, createTheme } from "@nextui-org/react";
+import Routes from "./routes";
 
-
-
+const theme = createTheme({
+  type: "dark", // it could be "light" or "dark"
+  theme: {
+    colors: {
+      // brand colors
+      // primaryLight: "$green200",
+      // primaryLightHover: "$green300",
+      // primaryLightActive: "$green400",
+      // primaryLightContrast: "$green600",
+      // primary: "#4ADE7B",
+      // primaryBorder: "$green500",
+      // primaryBorderHover: "$green600",
+      // primarySolidHover: "$green700",
+      // primarySolidContrast: "$white",
+      // primaryShadow: "$green500",
+      // gradient:
+      //   "linear-gradient(112deg, $blue100 -25%, $pink500 -10%, $purple500 80%)",
+      // link: "#5E1DAD",
+      // you can also create your own color
+      // myColor: "#ff4ecd",
+      // ...  more colors
+    },
+    space: {},
+    fonts: {},
+  },
+});
 export default function Home() {
   return (
-    
-    <Card5/>
-  )
+    <>
+      <NextUIProvider theme={theme}>
+        <BrowserRouter>
+          <Routes />
+        </BrowserRouter>
+      </NextUIProvider>
+    </>
+  );
 }
